+++ src/engine/dccnit.h
@@ -234,7 +234,7 @@ class engine
        * \param screen -> Pointer to the Screen's Surface
        * \param actualFarView -> FarView Value
        ***************************************************************/
-      void redefineWindow(SDL_Surface *screen, int actualFarView);
+      void redefineWindow(SDL_Surface *screen, float actualFarView);
 
       /*!
        ***************************************************************
